added controllers and routes for youtube video infro and lyrics display.  currently using temporary search params
