+++ routes/apiRoutes/lyricsRoutes/index.js
@@ -1,5 +1,5 @@
 const router            = require('express').Router();
-const lyricsController    = require('./../../../controllers/lyricsController');
+const searchController    = require('../../../controllers/searchController');
 
 
 // api/blogs prepended to every route declared in here
@@ -9,6 +9,12 @@ const authMiddlewares = require('./../../../middlewares/authMiddlewares');
 
 // /api/blogs
 router.route('/search')
-    .get(lyricsController.getSongs)
+    .get(searchController.getSongs)
+
+router.route('/display')
+    .get(searchController.getLyrics)
+
+router.route('/videoInfo')
+    .get(searchController.getVideoInfo)
 
 module.exports = router
